\name{dcTensor-package}
\alias{dcTensor-package}
\alias{dcTensor}
\docType{package}
\title{
\packageTitle{dcTensor}
}
\description{
\packageDescription{dcTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dcTensor}
\packageIndices{dcTensor}
}
\author{
\packageAuthor{dcTensor}

Maintainer: \packageMaintainer{dcTensor}
}
\references{
XXXXXXXXXXXX
}

\keyword{ package }
\seealso{\code{\link{toyModel}},\code{\link{dNMF}},\code{\link{dGDSVD}}}
\examples{
ls("package:dcTensor")
}