\name{dcTensor-package}
\alias{dcTensor-package}
\alias{dcTensor}
\docType{package}
\title{
\packageTitle{dcTensor}
}
\description{
\packageDescription{dcTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dcTensor}
\packageIndices{dcTensor}
}
\author{
\packageAuthor{dcTensor}

Maintainer: \packageMaintainer{dcTensor}
}
\references{
Z. Zhang, T. Li, C. Ding and X. Zhang, (2007). Binary Matrix Factorization with Applications, \emph{Seventh IEEE International Conference on Data Mining (ICDM 2007)}, 391-400
}

\keyword{ package }
\seealso{\code{\link{toyModel}},\code{\link{dNMF}},\code{\link{dGDSVD}}}
\examples{
ls("package:dcTensor")
}